cicd-evpn/requirements.txt
ansible==2.6.3
ansible-runner==1.2.0
appnope==0.1.0
asn1crypto==0.24.0
backcall==0.1.0
backports.ssl-match-hostname==3.5.0.1
bcrypt==3.1.4
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
cryptography==2.3.1
decorator==4.3.0
dill==0.2.8.2
docker-py==1.10.6
docker-pycreds==0.3.0
docutils==0.14
enum34==1.1.6
future==0.17.1
genie==3.1.0
genie.abstract==3.1.0
genie.conf==3.1.0
genie.examples==3.1.1
genie.harness==3.1.3
genie.libs.conf==3.1.2
genie.libs.filetransferutils==3.1.0
genie.libs.ops==3.1.3
genie.libs.parser==3.1.12
genie.libs.robot==3.1.4
genie.libs.sdk==3.1.5
genie.libs.telemetry==3.1.2
genie.metaparser==3.1.0
genie.ops==3.1.1
genie.parsergen==3.1.1
genie.predcore==3.1.0
genie.telemetry==3.1.4
genie.utils==3.1.0
idna==2.7
ipaddress==1.0.22
ipyats==0.2.0
ipython==7.1.1
ipython-genutils==0.2.0
jedi==0.13.1
Jinja2==2.10
jsonpickle==1.0
junit-xml==1.8
lockfile==0.12.2
MarkupSafe==1.0
netaddr==0.7.19
paramiko==2.4.1
parso==0.3.1
pexpect==4.6.0
pickleshare==0.7.5
prompt-toolkit==2.0.7
psutil==5.4.8
ptyprocess==0.6.0
pyasn1==0.4.4
pyats==5.0.1
pyats.aereport==5.0.1
pyats.aetest==5.0.0
pyats.async==5.0.0
pyats.connections==5.0.0
pyats.datastructures==5.0.0
pyats.easypy==5.0.2
pyats.examples==5.0.0
pyats.kleenex==5.0.2
pyats.log==5.0.0
pyats.results==5.0.0
pyats.robot==5.0.0
pyats.tcl==5.0.0
pyats.templates==5.0.0
pyats.topology==5.0.0
pyats.utils==5.0.3
pycparser==2.18
Pygments==2.2.0
PyNaCl==1.2.1
python-daemon==2.2.0
PyYAML==3.13
requests==2.19.1
requests-toolbelt==0.8.0
robotframework==3.0.4
setproctitle==1.1.10
six==1.11.0
tabulate==0.8.2
traitlets==4.3.2
unicon==3.4.0
urllib3==1.23
wcwidth==0.1.7
webexteamssdk==1.1.1
websocket-client==0.51.0
xmltodict==0.11.0
